pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "ml_core"
version = "0.1.0"
dependencies = [
    # TODO: Migrate dependencies from requirements.txt here
    "numpy",
    "torch",
    "torchvision",
    "h5py",
    "pyyaml",
    "tqdm",
    "torcheval",
    "tensorboard",
    "matplotlib",
    "seaborn",
    "pandas",
    "pytest",
    "ruff",
    "pre-commit",
    "wandb",
    "hydra-core",
    "omegaconf",
    "scikit-learn"
]

[tool.ruff]
line-length = 88
exclude = [".git", "__pycache__", "venv"]

[tool.ruff.lint]
#"I" handles import sorting, "B" finds likely bugs
select = ["E", "F", "I", "B"]
ignore = ["E501"] # Ignore long lines, we let the formatter handle it

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.poetry.dependencies]
hydra-core = "^1.3.2"